nurse306_hospital -- add getting nurse's order, patient, qualification files

+++ admin_backend/src/main/java/com/cooltoo/admin/api/NurseQualificationManageAPI.java
@@ -106,7 +106,7 @@ public class NurseQualificationManageAPI {
     @Produces(MediaType.APPLICATION_JSON)
     @AdminUserLoginAuthentication(requireUserLogin = true)
     public Response updateNurseQualification(@Context HttpServletRequest request,
-                                             @FormDataParam("qualification_file_id") int fileId,
+                                             @FormDataParam("qualification_file_id") long fileId,
                                              @FormDataParam("type") String workfileType,
                                              @FormDataParam("expiry") String expiry,
                                              @FormDataParam("file_name") String fileName,
